move CountAggregationResult implementation into own file
