+++ searchlib/src/vespa/searchlib/aggregation/countaggregationresult.h
@@ -7,22 +7,27 @@
 
 namespace search::aggregation {
 
+/**
+ * Counting aggregator that counts every single-value hit as 1, and every
+ * multi-value hit as its size.
+ */
 class CountAggregationResult : public AggregationResult {
+    expression::Int64ResultNode _count;
+
 public:
     DECLARE_AGGREGATIONRESULT(CountAggregationResult);
-    CountAggregationResult(uint64_t count = 0) : AggregationResult(), _count(count) {}
+    CountAggregationResult(uint64_t count = 0);
     void visitMembers(vespalib::ObjectVisitor& visitor) const override;
-    uint64_t getCount() const { return _count.get(); }
+    [[nodiscard]] uint64_t getCount() const { return _count.get(); }
     CountAggregationResult& setCount(uint64_t c) {
         _count = c;
         return *this;
     }
 
 private:
-    const ResultNode& onGetRank() const override { return _count; }
+    [[nodiscard]] const ResultNode& onGetRank() const override { return _count; }
     void onPrepare(const ResultNode& result) override;
     void initForUnitTest(const ResultNode& result) override;
-    expression::Int64ResultNode _count;
 };
 
 } // namespace search::aggregation
